Remove need for attach by user, add status to pipeline run, and clean up demo.
